phases/PhaseII/backend/requirements.txt
fastapi==0.115.0
sqlmodel==0.0.22
pyjwt==2.9.0
python-multipart==0.0.12
uvicorn[standard]==0.32.0
psycopg2-binary==2.9.10
pydantic-settings==2.6.1